% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{answer_questions}
\alias{answer_questions}
\title{Provide answers to various questions}
\usage{
answer_questions(dataset = NULL, multiple_disconnected = NULL,
  expect_topology = NULL, expected_topology = NULL,
  expect_cycles = NULL, expect_complex_tree = NULL,
  prior_information = NULL, n_cells = "< 10000",
  n_features = "< 1000", metric_importance = c(0.25, 0.25, 0.25, 0.25),
  dynmethods = TRUE, docker = TRUE, programming_interface = TRUE,
  languages = c("python", "C++", "R"), user_friendliness = 60,
  running_time = 5, n_methods = 4)
}
\arguments{
\item{dataset}{The dataset from which the answers will be computed}
}
\description{
Provide answers to various questions
}
