% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{answers}
\alias{answers}
\title{Provide answers to various questions}
\usage{
answers(multiple_disconnected = character(0),
  expect_topology = character(0), expected_topology = character(0),
  expect_cycles = character(0), expect_complex_tree = character(0),
  prior_information = NULL, n_cells = "< 10000", n_features = "< 1000",
  dynmethods = "Yes", docker = "Yes", programming_interface = "Yes",
  languages = c("python", "R", "C++"), user_friendliness = 60,
  running_time = 5, n_methods = 4)
}
\description{
Provide answers to various questions
}
@param multiple_disconnected Do you expect multiple disconnected trajectories in the data? Defaults to character(0)
@param expect_topology Do you expect a particular topology in the data? Defaults to character(0)
@param expected_topology What is the expected topology Defaults to character(0)
@param expect_cycles Do you expect cycles in the data? Defaults to character(0)
@param expect_complex_tree Do you expect a complex tree in the data? Defaults to character(0)
@param prior_information Are you willing to provide the following prior information? Defaults to NULL
@param n_cells Number of cells Defaults to < 10000
@param n_features Number of features (genes) Defaults to < 1000
@param dynmethods Do you use dynmethods to run the methods? Defaults to Yes
@param docker Is docker installed? Defaults to Yes
@param programming_interface Can you work in a programming interface? Defaults to Yes
@param languages Which languages can you work with? Defaults to c("python", "R", "C++")
@param user_friendliness Minimal user friendliness score Defaults to 60
@param running_time Maximal estimated running time (minutes) Defaults to 5
@param n_methods Number of methods Defaults to 4
