% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{get_answers_code}
\alias{get_answers_code}
\title{Produces the code necessary to reproduce a particular set of guidelines}
\usage{
get_answers_code(answers = answer_questions())
}
\arguments{
\item{answers}{An answers tibble as generated by [answer_questions()]}
}
\description{
Produces the code necessary to reproduce a particular set of guidelines
}
