% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_ui.R
\name{get_guidelines_methods_table}
\alias{get_guidelines_methods_table}
\title{get_guidelines_methods_table(guidelines(), list(selected = "false"))}
\usage{
get_guidelines_methods_table(guidelines, show_columns = character())
}
\description{
get_guidelines_methods_table(guidelines(), list(selected = "false"))
}
