% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{get_questions}
\alias{get_questions}
\alias{get_question}
\title{Load in the questions}
\usage{
get_questions()

get_question(question_id)
}
\arguments{
\item{question_id}{The id of the questions}
}
\description{
Load in the questions
}
