% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidelines.R, R/guidelines_prompt.R,
%   R/guidelines_shiny.R
\name{guidelines}
\alias{guidelines}
\alias{guidelines_prompt}
\alias{guidelines_shiny}
\title{Select the top methods, optionally based on a given task}
\usage{
guidelines(task = NULL, answers = list())

guidelines_prompt(task = NULL, answers = list())

guidelines_shiny(task = NULL, answers = list())
}
\arguments{
\item{task}{The task, optional}

\item{answers}{Optional, pre-provided answers to the different questions}
}
\description{
Select the top methods, optionally based on a given task
}
