% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R, R/guidelines.R, R/prompt.R
\name{guidelines_shiny}
\alias{guidelines_shiny}
\alias{guidelines}
\alias{guidelines_prompt}
\title{Select the top methods, optionally based on a given task}
\usage{
guidelines_shiny(task = NULL, answers = NULL)

guidelines(task = NULL, answers = list(programming_interface = "Yes",
  prior_information = c("Start cell"), n_methods = 4),
  method_columns = tibble(column_id = c("selected", "method_name",
  "topology_inference_type", "maximal_trajectory_type", "overall_benchmark")))

guidelines_prompt(task = NULL, answers = list())
}
\arguments{
\item{task}{The task, optional}

\item{answers}{Optional, pre-provided answers to the different questions}

\item{method_columns}{The columns to return}
}
\description{
Select the top methods, optionally based on a given task
}
