% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{data}
\name{priors}
\alias{priors}
\title{The prior information}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 3 columns.}
\usage{
priors
}
\description{
The prior information
}
\keyword{datasets}
