% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_server.R
\name{shiny_server}
\alias{shiny_server}
\title{The shiny server}
\usage{
shiny_server(answers = answer_questions())
}
\arguments{
\item{answers}{Previous answers other than default, see the \code{\link[=answer_questions]{answer_questions()}} function}
}
\description{
The shiny server
}
