% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{answer_questions}
\alias{answer_questions}
\title{Provide answers to various questions}
\usage{
answer_questions(dataset = NULL, ...)
}
\arguments{
\item{dataset}{The dataset from which the answers will be computed}

\item{...}{Answers to questions:
\itemize{
\item multiple_disconnected: Do you expect multiple disconnected trajectories in the data?, defaults to NULL:
\item expect_topology: Do you expect a particular topology in the data?, defaults to NULL:
\item expected_topology: What is the expected topology, defaults to NULL:
\item expect_cycles: Do you expect cycles in the data?, defaults to NULL:
\item expect_complex_tree: Do you expect a complex tree in the data?, defaults to NULL:
\item n_cells: Number of cells, defaults to 1000:
\item n_features: Number of features (genes), defaults to 1000:
\item time: Time limit, defaults to 1h:
\item memory: Memory limit, defaults to 2GB:
\item prior_information: Are you able to provide the following prior information?, defaults to NULL:
\item method_selection: How to select the number of methods, defaults to dynamic_n_methods:
\item dynamic_n_methods: Minimal probability of selecting the top model for the task, defaults to 80:
\item fixed_n_methods: Number of methods, defaults to 4:
\item metric_importance: How important are the following aspects of the trajectory?, defaults to list(correlation = 0.25, him = 0.25, featureimp_wcor = 0.25, F1_branches = 0.25):
\item user: Are you an end-user or a method developer?, defaults to user:
\item dynmethods: Do you use dynmethods to run the methods?, defaults to TRUE:
\item docker: Is docker installed?, defaults to TRUE:
\item programming_interface: Can you work in a programming interface?, defaults to TRUE:
\item languages: Which languages can you work with?, defaults to c("python", "R", "C++"):
\item user_friendliness: Minimal user friendliness score, defaults to 60:
\item developer_friendliness: Minimal developer friendliness score, defaults to 60:
\item exclude_datasets: Which datasets should be excluded, defaults to character(0):
}}
}
\description{
Provide answers to various questions
}
