% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{answer_questions}
\alias{answer_questions}
\title{Provide answers to various questions}
\usage{
answer_questions(dataset = NULL, multiple_disconnected = NULL,
  expect_topology = NULL, expected_topology = NULL,
  expect_cycles = NULL, expect_complex_tree = NULL, n_cells = 1000,
  n_features = 1000, time = "1h", memory = "2GB",
  prior_information = NULL, method_selection = "dynamic_n_methods",
  dynamic_n_methods = 80, fixed_n_methods = 4,
  metric_importance = c(0.25, 0.25, 0.25, 0.25), dynmethods = TRUE,
  docker = dynwrap::test_docker_installation(),
  programming_interface = TRUE, languages = c("python", "R", "C++"),
  user_friendliness = 60)
}
\arguments{
\item{dataset}{The dataset from which the answers will be computed}

\item{multiple_disconnected}{Do you expect multiple disconnected trajectories in the data? Defaults to NULL}

\item{expect_topology}{Do you expect a particular topology in the data? Defaults to NULL}

\item{expected_topology}{What is the expected topology Defaults to NULL}

\item{expect_cycles}{Do you expect cycles in the data? Defaults to NULL}

\item{expect_complex_tree}{Do you expect a complex tree in the data? Defaults to NULL}

\item{n_cells}{Number of cells Defaults to 1000}

\item{n_features}{Number of features (genes) Defaults to 1000}

\item{time}{Time limit Defaults to 1h}

\item{memory}{Memory limit Defaults to 2GB}

\item{prior_information}{Are you able to provide the following prior information? Defaults to NULL}

\item{method_selection}{How to select the number of methods Defaults to dynamic_n_methods}

\item{dynamic_n_methods}{Minimal probability of selecting the top model for the task Defaults to 80}

\item{fixed_n_methods}{Number of methods Defaults to 4}

\item{metric_importance}{How important are the following aspects of the trajectory? Defaults to c(0.25, 0.25, 0.25, 0.25)}

\item{dynmethods}{Do you use dynmethods to run the methods? Defaults to TRUE}

\item{docker}{Is docker installed? Defaults to dynwrap::test_docker_installation()}

\item{programming_interface}{Can you work in a programming interface? Defaults to TRUE}

\item{languages}{Which languages can you work with? Defaults to c("python", "R", "C++")}

\item{user_friendliness}{Minimal user friendliness score Defaults to 60}
}
\description{
Provide answers to various questions
}
