% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{answer_questions}
\alias{answer_questions}
\title{Provide answers to various questions}
\usage{
answer_questions(multiple_disconnected = TRUE, expect_topology = NULL,
  expected_topology = NULL, expect_cycles = NULL,
  expect_complex_tree = NULL, prior_information = NULL,
  n_cells = "< 10000", n_features = "< 1000", dynmethods = "Yes",
  docker = "Yes", programming_interface = "Yes",
  languages = c("python", "C++", "R"), user_friendliness = 60,
  running_time = 5, n_methods = 4)
}
\arguments{
\item{multiple_disconnected}{Do you expect multiple disconnected trajectories in the data? Defaults to TRUE}

\item{expect_topology}{Do you expect a particular topology in the data? Defaults to NULL}

\item{expected_topology}{What is the expected topology Defaults to NULL}

\item{expect_cycles}{Do you expect cycles in the data? Defaults to NULL}

\item{expect_complex_tree}{Do you expect a complex tree in the data? Defaults to NULL}

\item{prior_information}{Are you willing to provide the following prior information? Defaults to NULL}

\item{n_cells}{Number of cells Defaults to < 10000}

\item{n_features}{Number of features (genes) Defaults to < 1000}

\item{dynmethods}{Do you use dynmethods to run the methods? Defaults to Yes}

\item{docker}{Is docker installed? Defaults to Yes}

\item{programming_interface}{Can you work in a programming interface? Defaults to Yes}

\item{languages}{Which languages can you work with? Defaults to c("python", "C++", "R")}

\item{user_friendliness}{Minimal user friendliness score Defaults to 60}

\item{running_time}{Maximal estimated running time (minutes) Defaults to 5}

\item{n_methods}{Number of methods Defaults to 4}
}
\description{
Provide answers to various questions
}
