% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answers.R
\name{answers}
\alias{answers}
\title{Provide answers to various questions}
\usage{
answers()
}
\arguments{
\item{multiple_disconnected}{Do you expect multiple disconnected trajectories in the data? Defaults to}

\item{expect_topology}{Do you expect a particular topology in the data? Defaults to}

\item{expected_topology}{What is the expected topology Defaults to}

\item{expect_cycles}{Do you expect cycles in the data? Defaults to}

\item{expect_complex_tree}{Do you expect a complex tree in the data? Defaults to}

\item{prior_information}{Are you willing to provide the following prior information? Defaults to}

\item{n_cells}{Number of cells Defaults to}

\item{n_features}{Number of features (genes) Defaults to}

\item{metric_importance}{How important are the following aspects of the trajectory? Defaults to}

\item{dynmethods}{Do you use dynmethods to run the methods? Defaults to}

\item{docker}{Is docker installed? Defaults to}

\item{programming_interface}{Can you work in a programming interface? Defaults to}

\item{languages}{Which languages can you work with? Defaults to}

\item{user_friendliness}{Minimal user friendliness score Defaults to}

\item{running_time}{Maximal estimated running time (minutes) Defaults to}

\item{n_methods}{Number of methods Defaults to}
}
\description{
Provide answers to various questions
}
