% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{get_questions}
\alias{get_questions}
\alias{get_question}
\title{Load in the questions}
\usage{
get_questions(question_categories, answers)

get_question(question_id)
}
\description{
Load in the questions
}
