% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidelines.R, R/guidelines_shiny.R
\name{guidelines}
\alias{guidelines}
\alias{guidelines_shiny}
\title{Select the top methods, optionally based on a given dataset}
\usage{
guidelines(dataset = NULL, answers = answer_questions(dataset =
  dataset))

guidelines_shiny(dataset = NULL, answers = answer_questions(dataset =
  dataset))
}
\arguments{
\item{dataset}{The dataset, optional}

\item{answers}{Optional, pre-provided answers to the different questions. See [answer_questions()]}
}
\value{
Returns a dynguidelines::guidelines object, containing
  - `methods`: Ordered tibble containing information about the selected methods
  - `method_columns`: Information about what columns in methods are given and whether the were used for filtering or ordering
  - `answers`: An answers object, can be further modified.
  - `methods_selected`: Identifiers for all selected methods
}
\description{
Select the top methods, optionally based on a given dataset
}
